+++ include/nc_runtime/nc_types.h
@@ -12,6 +12,8 @@
 #include <stdio.h>
 #include <math.h>
 #include <float.h>
+
+#include <cmath>
 #include <new>
 
 #define UNUSED_VAR(o) (void)(o)